Dropping skips disabled keys
